Enable regular expression support based on whether UTF-8 is in the current locale (#5776)
* Regular expression support handling via UTF-8 in the locale


* Fixup some tests, including a typo in transpiler unicode fuzz test


* Update fuzz tests to not include \b or \B in fuzz testing because of
known issues with unicode


* Fix issue in fuzz tests with \Z followed by $


* Fix issue with word boundaries and negative character classes \D,\W,\S


* Add reference to issue regarding \b and \B unicode issue


* Fall back to CPU when negated character class is next to word boundary


* Add \H and \V to fallback scenario with word boundaries


* remove this test since it was removed in the upstream branch


* move word boundary fuzz testing logic to a separate flag skipUnicodeIssues which will skip when testing full unicode characters but will use when using a smaller ASCII subset


* Update the jenkins scripts here to set the locale


* need to export LC_ALL in mvn_verify stage here


* add comment for LC_ALL


* Regexp compatibility doc update


* Update scalatests and premerge build script


* update build scripts to test regexp separately from other tests because of locale requirement


* Feedback: code cleanup


* Fix syntax errors in RegularExpressionSuite that prevent it from loading in non-UTF-8 environments


* register custom regexp mark


* updates to build script and test script


* revert the nightly build script updates
