+++ integration_tests/pytest.ini
@@ -30,5 +30,6 @@ markers =
     nightly_host_mem_consuming_case: case in nightly_resource_consuming_test that consume much more host memory than normal cases
     fuzz_test: Mark fuzz tests
     iceberg: Mark a test that requires Iceberg has been configured, skipping if tests are not configured for Iceberg
+    regexp: Mark a test that tests regular expressions on the GPU (only works when UTF-8 is enabled)
 filterwarnings =
     ignore:.*pytest.mark.order.*:_pytest.warning_types.PytestUnknownMarkWarning
