+++ tests/src/test/scala/com/nvidia/spark/rapids/RegularExpressionSuite.scala
@@ -15,6 +15,8 @@
  */
 package com.nvidia.spark.rapids
 
+import java.nio.charset.Charset
+
 import com.nvidia.spark.rapids.shims.SparkShimImpl
 
 import org.apache.spark.SparkConf
@@ -60,33 +62,39 @@ class RegularExpressionSuite extends SparkQueryCompareTestSuite {
   }
 
   testSparkResultsAreEqual("String regexp_replace regex 1",
-    nullableStringsFromCsv, conf = conf) {
-    frame => frame.selectExpr("regexp_replace(strings,'.*','D')")
+    nullableStringsFromCsv, conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_replace(strings,'.*','D')")
   }
 
   testSparkResultsAreEqual("String regexp_replace regex 2",
-    nullableStringsFromCsv, conf = conf) {
-    frame => frame.selectExpr("regexp_replace(strings,'[a-z]+','D')")
+    nullableStringsFromCsv, conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_replace(strings,'[a-z]+','D')")
   }
 
   testSparkResultsAreEqual("String regexp_replace regex 3",
-    nullableStringsFromCsv,  conf = conf) {
-    frame => frame.selectExpr("regexp_replace(strings,'foo$','D')")
+    nullableStringsFromCsv,  conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_replace(strings,'foo$','D')")
   }
 
   testSparkResultsAreEqual("String regexp_replace regex 4",
-    nullableStringsFromCsv, conf = conf) {
-    frame => frame.selectExpr("regexp_replace(strings,'^foo','D')")
+    nullableStringsFromCsv, conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_replace(strings,'^foo','D')")
   }
 
   testSparkResultsAreEqual("String regexp_replace regex 5",
-    nullableStringsFromCsv, conf = conf) {
-    frame => frame.selectExpr("regexp_replace(strings,'(foo)','D')")
+    nullableStringsFromCsv, conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_replace(strings,'(foo)','D')")
   }
 
   testSparkResultsAreEqual("String regexp_replace regex 6",
-    nullableStringsFromCsv, conf = conf) {
-    frame => frame.selectExpr("regexp_replace(strings,'\\(foo\\)','D')")
+    nullableStringsFromCsv, conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_replace(strings,'\\(foo\\)','D')")
   }
 
   // https://github.com/NVIDIA/spark-rapids/issues/5659
@@ -106,8 +114,9 @@ class RegularExpressionSuite extends SparkQueryCompareTestSuite {
   // note that regexp_extract with a literal string gets replaced with the literal result of
   // the regexp_extract call on CPU
   testSparkResultsAreEqual("String regexp_extract literal input",
-    extractStrings, conf = conf) {
-    frame => frame.selectExpr("regexp_extract('abc123def', '^([a-z]*)([0-9]*)([a-z]*)$', 2)")
+    extractStrings, conf = conf) { frame =>
+      assume(isUnicodeEnabled())
+      frame.selectExpr("regexp_extract('abc123def', '^([a-z]*)([0-9]*)([a-z]*)$', 2)")
   }
 
   private def extractStrings(session: SparkSession): DataFrame = {
@@ -119,6 +128,10 @@ class RegularExpressionSuite extends SparkQueryCompareTestSuite {
       ("abc\r\n12\r3\ndef"),
       ("123abc456")
     ).toDF("strings")
+  }
+
+  private def isUnicodeEnabled(): Boolean = {
+    Charset.defaultCharset().name() == "UTF-8"
   }
 
 }
